take out NewTenant
